🩹 Fixed issue with circular imports
